The BridgeSampling Module

BridgeSampling.bridgesampling — Function

bridgesampling(samples, log_posterior, lb, ub; tol, maxiter, names)

Estimates the log marginal likelihood with bridge sampling, given a matrix of posterior samples and an unnormalized log posterior function.

# Arguments
* `samples` : An ndims x nsamples matrix. The eltype can be either a scalar or an array of any dimension (e.g. multivariate parameters)
* `log_posterior` : A function that takes one argument (one sample) and outputs the unnormalized log posterior probability (log joint)
* `lb` and `ub` : respectively the lower and upper boundary of each dimension in the samples
* `tol` : tolerance for the iterative scheme. Default = `1e-10`
* `maxiter` : maximum number of iterations of the iterative scheme. Default = `1000`
* `names` : A vector of names for each dimension in the samples. If specified, the input argument of `log_posterior` will be a `NamedTuple`. Default = `nothing`

# Examples
```julia 
    # Bivariate standard normal distribution
    samples = rand(MvNormal(ones(2)), 10_000)

    log_posterior(x) = -0.5 * x' * x

    Bridge = bridgesampling(samples, log_posterior, [-Inf, -Inf], [Inf, Inf])
```

# Sources 
    Quentin F. Gronau, Alexandra Sarafoglou, Dora Matzke, Alexander Ly, Udo Boehm, Maarten Marsman, David S. Leslie, Jonathan J. Forster, Eric-Jan Wagenmakers, Helen Steingroever,
    A tutorial on bridge sampling,
    Journal of Mathematical Psychology,
    Volume 81,
    2017,
    Pages 80-97,
    ISSN 0022-2496,
    https://doi.org/10.1016/j.jmp.2017.09.005.

Missing docstring.

Missing docstring for ````

Module Index`. Check Documenter's build log for details.

@index Modules = [BridgeSampling] Order = [:constant, :type, :function, :macro]

## Detailed API

@autodocs Modules = [BrdigeSampling] Order = [:constant, :type, :function, :macro] ```